+++ front/src/components/BackButton/styles.ts
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+const ArrowIcon = styled.img`
+  position: absolute;
+  left: -1em;
+  bottom: 1em;
+  width: 50px;
+  height: 50px;
+  padding: 10px;
+  border-radius: 50%;
+  background-color: ${({ theme }) => theme.colors.BLUE_800};
+  cursor: pointer;
+
+  :hover {
+    opacity: 0.7;
+    transform: scale(1.1);
+    transition: ease-in-out 0.2s;
+  }
+`;
+
+export { ArrowIcon };
