+++ front/src/components/HistoryItem/styles.ts
@@ -4,13 +4,13 @@ const Item = styled.div<{ isSelected: boolean }>`
   margin: 0 16px 10px 0;
   padding: 8px 12px;
   background-color: ${({ theme, isSelected }) =>
-    isSelected ? theme.colors.GREEN_400 : theme.colors.GREEN_100};
+    isSelected ? theme.colors.GRAY_200 : theme.colors.GRAY_100};
   border-radius: 8px;
   border: 1px solid ${({ theme }) => theme.colors.GRAY_100};
   cursor: pointer;
 
   :hover {
-    border: 1px solid ${({ theme }) => theme.colors.GRAY_500};
+    border: 1px solid ${({ theme }) => theme.colors.GRAY_300};
   }
 `;
