+++ front/src/components/ReservationInfo/index.tsx
@@ -1,19 +1,18 @@
+import BackButton from '@components/BackButton';
 import { getHourMinutes, getMonthDate } from '@utils/index';
 import * as S from './styles';
 
 import ScheduleIcon from '@assets/schedule.svg';
 import ClockIcon from '@assets/clock.svg';
-import LeftArrowIcon from '@assets/left-arrow-disabled.svg';
 
 interface ReservationInfoProps {
   image?: string;
   name?: string;
   dateTime?: string;
   isView: boolean;
-  onClick: () => void;
 }
 
-const ReservationInfo = ({ image, name, dateTime, isView, onClick }: ReservationInfoProps) => {
+const ReservationInfo = ({ image, name, dateTime, isView }: ReservationInfoProps) => {
   return (
     <>
       <S.CoachImg src={image} alt="코치 프로필 이미지" />
@@ -26,7 +25,7 @@ const ReservationInfo = ({ image, name, dateTime, isView, onClick }: Reservation
         <img src={ClockIcon} alt="시계 아이콘" />
         <span>{getHourMinutes(dateTime as string)}</span>
       </S.DateWrapper>
-      {isView && <S.ArrowIcon src={LeftArrowIcon} alt="화살표 아이콘" onClick={onClick} />}
+      {isView && <BackButton />}
     </>
   );
 };
