+++ front/src/pages/CoachSheet/index.tsx
@@ -1,4 +1,4 @@
-import { useParams, useNavigate, useLocation } from 'react-router-dom';
+import { useParams, useLocation } from 'react-router-dom';
 
 import Frame from '@components/Frame';
 import ReservationInfo from '@components/ReservationInfo';
@@ -8,7 +8,6 @@ import { Reservation } from '@typings/domain';
 import * as S from '@styles/common';
 
 const CoachSheet = () => {
-  const navigate = useNavigate();
   const { id: reservationId } = useParams();
   const { state: crewId } = useLocation();
 
@@ -26,7 +25,6 @@ const CoachSheet = () => {
           name={reservationInfo.coachName}
           dateTime={reservationInfo.dateTime}
           isView={true}
-          onClick={() => navigate(-1)}
         />
       </S.InfoContainer>
       <Sheet title="작성한 면담 내용" sheets={reservationInfo.sheets} isView={true} />
